+++ terraform/slurm_cluster/variables.tf
@@ -172,6 +172,14 @@ EOD
 # LOGIN #
 #########
 
+variable "enable_login" {
+  description = <<EOD
+Enables the creation of login nodes and instance templates.
+EOD
+  type        = bool
+  default     = true
+}
+
 variable "login_nodes" {
   description = "List of slurm login instance definitions."
   type = list(object({